[dashboard] Inline benchmark report comparison on timeline
Click a point on the daily benchmark chart to mark it as the report, then
click an earlier point to set a baseline — a full comparison renders inline
below the chart using the already-loaded reports (no refetching). Role
assignment is date-aware: clicking a later point promotes it to report and
demotes the current report to baseline.

Chart-level changes:
- One line per test (totalDuration) instead of one per render/phase
- Selected commits shown as labeled ChartsReferenceLine verticals
- Baseline/Report chips under the chart
- Single shared LineChart for single vs multi-benchmark mode

The comparison UI was extracted from BenchmarkDetails into a pure
BenchmarkComparisonReportView that takes value/base reports and owns
compareBenchmarkReports internally.

Entry sort now uses a precomputed priority on ComparisonItem encoding
severity (error → success → neutral) and absolute diff magnitude in one
number, so the shared view always shows regressions first.

Also moves scrollbar-gutter: stable from body to html so the rule applies
to the real viewport scroll container and the benchmarks page no longer
shifts when the comparison section appears.
